requirements.txt
python-telegram-bot==13.1
pytest==6.1.2
pytest-mock==3.3.1